scripts/requirements.txt
pyyaml
pandas
gitpython
setuptools
ruamel.yaml>=0.15
